Refactor chip support and configuration
Remove 'family' from the layers. Each chip module now brings in the
specific drivers it supports. We move all of these kinds of drivers into
a drivers directory. The configuration module lives inline, within the
chip module.
