+++ src/chip/drivers/ccm_10xx/analog/pll6_500mhz.rs
