+++ src/chip/drivers/ccm_10xx/periph_clk2_sel.rs
