+++ src/chip/drivers/ccm_11xx/clock_gate.rs
