+++ src/chip/none.rs
@@ -4,7 +4,4 @@ pub mod ccm {
     pub use crate::common::ccm::*;
 }
 pub mod dma {}
-pub mod lpi2c {}
-
-pub(crate) mod reexports {}
 pub(crate) mod iomuxc {}
